pyproject.toml
[project]
name = "blog-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.35.0",
    "openai>=1.107.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-prometheus>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "pydantic>=2.11.7",
    "tiktoken>=0.11.0",
    "weaviate-client>=4.16.9",
]
